+++ src/mutation/create-user.use-case.ts
@@ -1,4 +1,10 @@
 import axios from "axios";
+import { hashPassword } from "../cryptography/password-encription";
+import { AppDataSource } from "../data-source";
+import { User } from "../entity/User";
+import { CustomError } from "../errors/login-error-class";
+import { isRepeatedEmail } from "../validators/email-validator";
+import { validatePassword } from "../validators/password-validator";
 
 export interface CreateUserInput {
   name: string;
@@ -7,24 +13,31 @@ export interface CreateUserInput {
   password: string;
 }
 
-export async function createUserMutation(input: CreateUserInput) {
-  const response = await axios.post("http://localhost:4001/graphql", {
-    operationName: null,
-    variables: {
-      data: input,
-    },
-    query: `mutation CreateUser($data: UserInput!) {
-        createUser(
-          data: $data
-        ) {
-          id
-          email
-          name
-          birthDate
-        }
-      }
-      `,
-  });
+export async function createUser(data: CreateUserInput) {
+  const validPassword = validatePassword(data.password);
+  const repeatedEmail = await isRepeatedEmail(data.email);
 
-  return response;
+  if (!validPassword) {
+    throw new CustomError(
+      400,
+      "The password must contain at least 6 characters, of which 1 must be a letter and 1 a digit"
+    );
+  }
+  if (repeatedEmail) {
+    throw new CustomError(
+      400,
+      "This email has already been registered by another user"
+    );
+  }
+
+  const hashedPassword = hashPassword(data.password);
+
+  const newUser = new User();
+  newUser.name = data.name;
+  newUser.birthDate = data.birthDate;
+  newUser.email = data.email;
+  newUser.password = hashedPassword.hashedPassword;
+  newUser.salt = hashedPassword.salt;
+  await AppDataSource.manager.save(newUser);
+  return newUser;
 }
